requirements.txt
tensorflow-cpu
tensorflow-hub
flask
gunicorn
Pillow
Flask-Table
flask-cors